requirements.txt
Flask
psycopg2-binary
python-dotenv
Flask-Cors
gunicorn
# whitenoise  # Uncomment if serving static files from Flask
